Add files via upload
Vulnerability Data Fetcher is our Flask-based web application designed to fetch, store, and display vulnerability data from the CVE Details API. The application allows users to configure the range of years for data fetching, dynamically fetches and paginates the data from the API, and stores it in a SQLite database. It includes features like dynamic API requests with 30-second intervals between yearly data fetching, database reset functionality, and an intuitive web interface for interacting with and visualizing stored records. The backend leverages SQLAlchemy for database management, while the frontend provides responsive and user-friendly features like table-based pagination. This project serves as a practical tool for cybersecurity professionals and researchers by streamlining vulnerability data collection and analysis. With a well-structured codebase and interactive design, it simplifies the process of handling large datasets and provides an efficient way to explore and analyze vulnerabilities. Built with Flask, SQLAlchemy, and JavaScript, the application is easily deployable and configurable, offering a comprehensive solution for vulnerability tracking and research.